requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-OAuthlib==0.9.6
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
psycopg2-binary==2.9.9
email-validator==2.1.0